Added support for fractional pricing

+++ src/main/java/com/andrei/supermarket/domain/Receipt.java
@@ -1,6 +1,7 @@
 package com.andrei.supermarket.domain;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record Receipt(List<ReceiptItem> items, int total) {
+public record Receipt(List<ReceiptItem> items, BigDecimal total) {
 }
